+++ cigvis/visernodes/volume_slice.py
@@ -4,6 +4,7 @@ import numpy as np
 from cigvis import colormap, ExceptionWrapper
 import matplotlib.pyplot as plt
 from cigvis.utils import utils
+from cigvis import is_line_first
 
 try:
     import viser
@@ -27,10 +28,27 @@ class VolumeSlice:
         self.axis = axis
         self.pos = pos
         self.cmap = cmap
-        self.clim = clim if clim is not None else utils.auto_clim()
+        self._cmap_preset = cmap
+        self.clim = clim if clim is not None else utils.auto_clim(volume)
 
         self.init_scale = [1.5 / max(volume.shape)] * 3
         self.nancolor = nancolor
+
+        # HACK: to deal with rgb or rgba image
+        def _eq_3_or_4(k):
+            return k == 3 or k == 4
+        line_first = is_line_first()
+        assert _eq_3_or_4(volume.ndim), f"Volume's dims must be 3 or 4 (RGB), but got {volume.ndim}"
+        # rgb_type, 0 for (n1, n2, n3), 1 for (n1, n2, n3, 3/4), 2 for (3/4, n1, n2, n3)
+        ndim = volume.ndim
+        self.channel_dim = None
+        dim_x, dim_y, dim_z = (0, 1, 2) if line_first else (2, 1, 0)
+        self.axis_to_dim = {'x': dim_x, 'y': dim_y, 'z': dim_z}
+        self.vol_shape, self.rgb_type = utils.get_shape(volume, line_first)
+        if self.rgb_type == 1:
+            self.channel_dim = 3
+        elif self.rgb_type == 2:
+            self.channel_dim = 0
 
         if isinstance(scale, (int, float)):
             if scale < 0:
@@ -45,6 +63,7 @@ class VolumeSlice:
         self.masks = []
         self.fg_cmaps = []
         self.fg_clims = []
+        self._fg_cmaps_preset = []
 
     @property
     def server(self):
@@ -60,16 +79,16 @@ class VolumeSlice:
     @property
     def render_width(self):
         if self.axis != 'z':
-            return self.volume.shape[2] * self.scale[2]
+            return self.vol_shape[2] * self.scale[2]
         else:
-            return self.volume.shape[1] * self.scale[1]
+            return self.vol_shape[1] * self.scale[1]
 
     @property
     def render_height(self):
         if self.axis == 'x':
-            return self.volume.shape[1] * self.scale[1]
+            return self.vol_shape[1] * self.scale[1]
         else:
-            return self.volume.shape[0] * self.scale[0]
+            return self.vol_shape[0] * self.scale[0]
 
     @property
     def wxyz(self):
@@ -82,7 +101,7 @@ class VolumeSlice:
 
     @property
     def position(self):
-        ni, nx, nt = self.volume.shape
+        ni, nx, nt = self.vol_shape
         ri = ni * self.scale[0]
         rx = nx * self.scale[1]
         rt = nt * self.scale[2]
@@ -98,16 +117,24 @@ class VolumeSlice:
             img = img.detach().cpu().numpy()
         return img
 
-    def to_img(self):
-        if self.axis == 'x':
-            bg = self.volume[self.pos, :, :]
-            fg = [mask[self.pos, :, :] for mask in self.masks]
-        elif self.axis == 'y':
-            bg = self.volume[:, self.pos, :]
-            fg = [mask[:, self.pos, :] for mask in self.masks]
+    def _get_slices(self, axis, pos):
+        dim = self.axis_to_dim.get(axis)
+        slices = [slice(None)] * self.volume.ndim
+        if self.channel_dim is not None and dim >= self.channel_dim:
+            slices[dim + 1] = pos
         else:
-            bg = self.volume[:, :, self.pos]
-            fg = [mask[:, :, self.pos] for mask in self.masks]
+            slices[dim] = pos
+        return tuple(slices)
+
+    def _auto_transpose(self, img):
+        if not is_line_first():
+            img = np.transpose(img, (1, 0, 2))
+        return img
+
+    def to_img(self):
+        s = self._get_slices(self.axis, self.pos)
+        bg = self.volume[s]
+        fg = [mask[s] for mask in self.masks]
 
         bg = self._to_np(bg)
         fg = [self._to_np(g) for g in fg]
@@ -115,7 +142,7 @@ class VolumeSlice:
         img = colormap.arrs_to_image([bg] + fg, [self.cmap] + self.fg_cmaps,
                                      [self.clim] + self.fg_clims, True,
                                      self.nancolor)
-        return img
+        return self._auto_transpose(img)
 
     def update_node(self, pos):
         if self.server is None:
@@ -130,7 +157,7 @@ class VolumeSlice:
             img,
             self.render_width,
             self.render_height,
-            'png',
+            format='png',
             wxyz=self.wxyz,
             position=self.position,
         )
@@ -142,11 +169,11 @@ class VolumeSlice:
     @property
     def limit(self):
         if self.axis == 'x':
-            return (0, self.volume.shape[0])
+            return (0, self.vol_shape[0])
         elif self.axis == 'y':
-            return (0, self.volume.shape[1])
+            return (0, self.vol_shape[1])
         else:
-            return (0, self.volume.shape[2])
+            return (0, self.vol_shape[2])
 
     def _check_bound(self):
         if self.pos < 0:
@@ -155,11 +182,25 @@ class VolumeSlice:
             self.pos = self.limit[1] - 1
 
     def update_cmap(self, cmap):
+        if cmap is None:
+            cmap = self._cmap_preset
         self.cmap = cmap
         self.update_node(self.pos)
 
     def update_clim(self, clim):
         self.clim = clim
+        self.update_node(self.pos)
+
+    def update_mask_clim(self, clim, num):
+        if len(self.masks) == 0 or num >= len(self.masks):
+            return 
+        self.fg_clims[num] = clim 
+        self.update_node(self.pos)
+
+    def update_mask_cmap(self, cmap, num):
+        if len(self.masks) == 0 or num >= len(self.masks):
+            return 
+        self.fg_cmaps[num] = cmap 
         self.update_node(self.pos)
 
     def update_scale(self, scale):
@@ -170,10 +211,12 @@ class VolumeSlice:
         self.update_node(self.pos)
 
     def add_mask(self, vol, cmap: str, clim: List = None):
-        assert vol.shape == self.volume.shape
+        mask_shape, _ = utils.get_shape(vol, is_line_first())
+        assert mask_shape == self.vol_shape, f"mask.shape: {vol.shape} != vol.shape: {self.vol_shape}"
         self.masks.append(vol)
         if clim is None:
             clim = utils.auto_clim(vol)
         self.fg_cmaps.append(cmap)
+        self._fg_cmaps_preset.append(cmap)
         self.fg_clims.append(clim)
         self.update_node(self.pos)